+++ packages/dual-channel/src/app/components/embedded-items/index.js
@@ -62,6 +62,7 @@ const itemSize = css`
 const Container = styled.div`
   overflow: hidden;
   position: relative;
+  background-color: #f1f1f1;
   ${mq.desktopOnly`
     width: ${mockup.itemWidth.desktop}px;
   `}
@@ -71,8 +72,10 @@ const Container = styled.div`
   ${mq.tabletBelow`
     z-index: ${styles.zIndex.embeddedItem};
     width: ${mockup.itemWidth.mobile};
-    ${props =>
-      props.sectionsPosition !== Waypoint.inside ? '' : `padding-bottom: 100%`};
+    height: ${props =>
+      props.sectionsPosition !== Waypoint.inside
+        ? '0'
+        : mockup.itemHeight.mobile};
     position: fixed;
     top: 0;
     left: 0;
@@ -86,6 +89,7 @@ const SectionWrapper = styled.div`
   ${mq.tabletBelow`
     left: 50%;
     transform: translateX(-50%);
+    max-width: ${mockup.itemHeight.mobile};
   `}
 `
 
@@ -131,7 +135,10 @@ const ItemAnimationWrapper = styled.div`
       }
       case 'none':
       default: {
-        return ''
+        return css`
+          opacity: ${props =>
+            props.isFocused || props.isPrevious ? '100' : '0'};
+        `
       }
     }
   }}
@@ -139,7 +146,21 @@ const ItemAnimationWrapper = styled.div`
   ${mq.tabletBelow`
     text-align: center;
     background-color: #f1f1f1;
+
+    /* default styles for img embedded items */
+    /* users can overwrite these styles in the spreadsheet */
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: contain;
+      object-position: center;
+    }
+
   `}
+
+  img {
+    width: 100%;
+  }
 `
 
 function SectionItem(props) {