+++ packages/dual-channel/src/app/components/root.js
@@ -72,7 +72,19 @@ const FirstEmbeddedItem = styled.div`
   ${mq.tabletBelow`
     height: 50vh;
     width: 100%;
+    max-width: 50vh;
     text-align: center;
+    margin-left: auto;
+    margin-right: auto;
+
+    /* default styles for img embedded items */
+    /* users can overwrite these styles in the spreadsheet */
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: contain;
+      object-position: center;
+    }
   `}
 
   ${mq.desktopOnly`
@@ -88,6 +100,10 @@ const FirstEmbeddedItem = styled.div`
   > div {
     width: 100%;
     height: 100%;
+  }
+
+  img {
+    width: 100%;
   }
 `