+++ packages/scrollable-image-ui/webpack.config.js
@@ -1,3 +1,4 @@
+// const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const path = require('path')
 const pkg = require('./package.json')
@@ -12,6 +13,7 @@ const config = {
   output: {
     path: path.resolve(__dirname, 'dist/'),
     filename: '[name].[hash].bundle.js',
+    chunkFilename: '[name].[hash].bundle.js',
     // For path translation, use URL provided by unpkg rather than relative path.
     //
     // The URL we want to make public is
@@ -28,7 +30,7 @@ const config = {
     // And it is incorrect URL.
     publicPath: isDevelopment
       ? '/'
-      : `https://unpkg.com/${pkg.name}@${pkg.version}/dist`,
+      : `https://unpkg.com/${pkg.name}@${pkg.version}/dist/`,
   },
   optimization: {
     minimize: !isDevelopment,
@@ -46,6 +48,44 @@ const config = {
             )
           },
           name: 'polyfill',
+          priority: 9,
+          enforce: true,
+        },
+        react: {
+          test: module => {
+            return (
+              module.context &&
+              /node_modules\/(react|history|redux|styled-components)/.test(
+                module.context
+              )
+            )
+          },
+          name: 'react-base',
+          priority: 10,
+          reuseExistingChunk: true,
+          enforce: true,
+        },
+        lodash: {
+          test: module => {
+            return (
+              module.context && module.context.includes('node_modules/lodash')
+            )
+          },
+          name: 'lodash',
+          priority: 11,
+          reuseExistingChunk: true,
+          enforce: true,
+        },
+        materialUI: {
+          test: module => {
+            return (
+              module.context &&
+              /node_modules\/@material-ui/.test(module.context)
+            )
+          },
+          name: 'material-ui',
+          priority: 8,
+          reuseExistingChunk: true,
           enforce: true,
         },
       },
@@ -76,6 +116,7 @@ const config = {
     new HtmlWebpackPlugin({
       template: path.join(__dirname, devDirname, 'index.html'),
     }),
+    // new BundleAnalyzerPlugin()
   ],
 }