+++ packages/timeline/src/components/unit-flag.js
@@ -1,4 +1,3 @@
-import { sourceHanSansTC as fontWeight } from '@twreporter/core/lib/constants/font-weight'
 import elementTypes from '../constants/element-types'
 import predefinedPropTypes from '../constants/prop-types'
 import React, { PureComponent } from 'react'
@@ -31,14 +30,14 @@ const Flag = styled.h3`
 
 const Label = styled.div`
   font-size: 16px;
-  font-weight: ${fontWeight.normal};
+  font-weight: ${props => props.theme[elementTypes.unitFlag].labelFontWeight};
   letter-spacing: 0.9px;
   flex: 0 0 auto;
 `
 
 const Title = styled.div`
   font-size: 16px;
-  font-weight: ${fontWeight.regular};
+  font-weight: ${props => props.theme[elementTypes.unitFlag].titleFontWeight};
   margin-left: 0.4em;
   flex: 0 1 auto;
 `