+++ packages/timeline/src/constants/default-theme.js
@@ -1,23 +1,36 @@
 import defaultFontFamily from './font-family'
 import elementTypes from './element-types'
+import { sourceHanSansTC as fontWeight } from '@twreporter/core/lib/constants/font-weight'
 
 export default {
   [elementTypes.record]: {
-    titleColor: '#a67a44',
     color: '#404040',
-    strongColor: '#262626',
+    figcaptionColor: '#808080',
+    figcaptionFontWeight: fontWeight.light,
+    fontFamily: defaultFontFamily,
     linkColor: '#a67a44',
     linkUnderlineColor: '#d8d8d8',
-    fontFamily: defaultFontFamily,
+    pFontWeight: fontWeight.normal,
+    strongColor: '#262626',
+    titleColor: '#a67a44',
+    titleFontWeight: fontWeight.medium,
   },
   [elementTypes.unitFlag]: {
     color: '#fff',
     background: '#000',
     fontFamily: defaultFontFamily,
+    labelFontWeight: fontWeight.normal,
+    titleFontWeight: fontWeight.regular,
   },
   [elementTypes.groupFlag]: {
     color: '#fff',
     background: '#a67a44',
     fontFamily: defaultFontFamily,
+    labelFontWeight: fontWeight.normal,
+    titleFontWeight: fontWeight.regular,
+  },
+  lineColor: '#000',
+  emphasizedElements: {
+    background: '#fff',
   },
 }